@@ -1,12 +1,11 @@
-import React from "react"
-import ReactDOM from "react-dom"
+import ReactDOM from "react-dom/client"
 import App from "./App"
 import "./App.scss"
 import { BrowserRouter } from "react-router-dom"
 
-ReactDOM.render(
+const root = ReactDOM.createRoot(document.getElementById("root"))
+root.render(
   <BrowserRouter>
     <App />
-  </BrowserRouter>,
-  document.getElementById("root")
+  </BrowserRouter>
 )
